requirements.txt
boto3>=1.34.0
botocore>=1.34.0
tabulate>=0.9.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0
